GDP-28 arreglar botones de editar
